moved files in components folder to be in folders named by the page they are used in
