+++ src/views/ViewAssessment.vue
@@ -24,11 +24,11 @@
 import { Component, Vue } from "vue-property-decorator";
 import { AssessmentComplete } from "@/models/assessment";
 
-import Renderer from "@/components/Renderer.vue";
-import ProjectInfoCard from "@/components/ProjectInfoCard.vue";
-import ABreakdownCard from "@/components/ABreakdownCard.vue";
-import MaterialBreakdownCard from "@/components/MaterialBreakdownCard.vue";
-import ViewAssessmentButtons from "@/components/ViewAssessmentButtons.vue";
+import Renderer from "@/components/shared/Renderer.vue";
+import ProjectInfoCard from "@/components/viewAssessment/ProjectInfoCard.vue";
+import ABreakdownCard from "@/components/viewAssessment/ABreakdownCard.vue";
+import MaterialBreakdownCard from "@/components/viewAssessment/MaterialBreakdownCard.vue";
+import ViewAssessmentButtons from "@/components/viewAssessment/ViewAssessmentButtons.vue";
 
 @Component({
   components: {
